변한버전 푼것 (완성) 문제 잘봐야함 입력값에 12승까지임

+++ src/test_11382/BigSum.java
@@ -12,10 +12,10 @@ public class BigSum {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 
         StringTokenizer st = new StringTokenizer(br.readLine());
-        int result = 0;
+        long result = 0;
 
         while (st.hasMoreTokens()) {
-            result += Integer.parseInt(st.nextToken());
+            result += Long.parseLong(st.nextToken());
         }
 
         System.out.println(result);
